Create modules folder, task and elements files
